# src/main/resources/static/swagger.yml
openapi: 3.0.4
info:
  title: itPlace API Document
  description: |-
    지도 기반 내 주변 LG유플러스 멤버십 혜택 정보 조회 서비스 'itPlace'의 API 명세서입니다.
    
    모든 API 응답은 표준 응답 포맷(ApiResponse)을 따릅니다.
  version: 1.0.0
servers:
  - url: https://user-api.itplace.click/api/v1
    description: 개발 서버

  - url: http://localhost:8080/api/v1
    description: 로컬 테스트 서버
...
paths:
...
  /favorites:
    post:
      tags:
        - Favorites
      summary: 즐겨찾기 등록
      description: 사용자의 즐겨찾기 목록에 제휴처를 추가합니다.
      operationId: addFavorite
      requestBody:
        description: 혜택 ID를 포함한 요청 객체
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FavoriteRequest'
      responses:
        '200':
          description: 즐겨찾기 등록 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: FAVORITE_ADD_SUCCESS
                status: OK
                message: 즐겨찾기 등록이 완료되었습니다
                data: null
                timestamp: '2025-07-04T14:06:54.631Z'

        '404':
          description: 사용자 또는 혜택을 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                userNotFound:
                  summary: 사용자 없음
                  value:
                    code: USER_NOT_FOUND
                    status: NOT_FOUND
                    message: 사용자를 찾을 수 없습니다.
                    data: null
                    timestamp: '2025-07-16T15:00:00.000Z'
                benefitNotFound:
                  summary: 혜택 없음
                  value:
                    code: BENEFIT_NOT_FOUND
                    status: NOT_FOUND
                    message: 존재하지 않는 혜택입니다.
                    data: null
                    timestamp: '2025-07-16T15:00:00.000Z'

        '409':
          description: 이미 즐겨찾기에 등록된 경우
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: FAVORITE_ALREADY_EXISTS
                status: CONFLICT
                message: 이미 즐겨찾기에 등록된 혜택입니다.
                data: null
                timestamp: '2025-07-16T15:00:00.000Z'

    delete:
      tags:
        - Favorites
      summary: 즐겨찾기 삭제
      description: 사용자 ID와 혜택 ID를 기반으로 즐겨찾기를 삭제합니다.
      operationId: removeFavorites
      requestBody:
        description: 혜택 ID를 포함한 요청 객체
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RemoveFavoritesRequest'
      responses:
        '200':
          description: 즐겨찾기 삭제 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: FAVORITE_DELETE_SUCCESS
                status: OK
                message: 즐겨찾기 삭제가 완료되었습니다.
                data: null
                timestamp: '2025-07-07T12:00:00Z'

        '404':
          description: 사용자 또는 혜택을 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                userNotFound:
                  summary: 사용자 없음
                  value:
                    code: USER_NOT_FOUND
                    status: NOT_FOUND
                    message: 사용자를 찾을 수 없습니다.
                    data: null
                    timestamp: '2025-07-16T15:00:00.000Z'
                benefitNotFound:
                  summary: 혜택 없음
                  value:
                    code: BENEFIT_NOT_FOUND
                    status: NOT_FOUND
                    message: 존재하지 않는 혜택입니다.
                    data: null
                    timestamp: '2025-07-16T15:00:00.000Z'

    get:
      tags:
        - Favorites
      summary: 즐겨찾기한 혜택 목록 조회 - 페이징, 필터링(VIP 콕, 기본 혜택)
      description: 사용자가 즐겨찾기한 혜택 목록을 보여줍니다.
      operationId: getFavorites
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
            default: null
          description: 카테고리 설정 없을 시 전체 조회
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
          description: 조회할 페이지 번호 (0부터 시작)
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 6
          description: 페이지당 항목 수 (기본 6개)
      responses:
        '200':
          description: 즐겨찾기 목록 조회 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PageResult'
              examples:
                success:
                  summary: 즐겨찾기 목록 예시
                  value:
                    code: FAVORITE_BENEFIT_SUCCESS
                    status: OK
                    message: 즐겨찾기 목록 조회에 성공했습니다.
                    data:
                      content:
                        - benefitId: 1
                          benefitName: CGV
                          partnerName: CGV
                          partnerImage: S3객체URL
                        - benefitId: 2
                          benefitName: GS25
                          partnerName: GS25
                          partnerImage: S3객체URL
                      totalElements: 2
                      totalPages: 1
                      currentPage: 0
                      hasNext: false
                    timestamp: '2025-07-04T15:30:00.000Z'
        '404':
          description: 사용자 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: USER_NOT_FOUND
                status: NOT_FOUND
                message: 사용자를 찾을 수 없습니다.
                data: null
                timestamp: '2025-07-16T15:00:00.000Z'
  /favorites/search:
    get:
      tags:
        - Favorites
      summary: 즐겨찾기 혜택 이름 검색
      description: 로그인된 사용자의 즐겨찾기 목록에서 혜택 이름으로 검색합니다.
      operationId: searchFavorites
      parameters:
        - name: keyword
          in: query
          required: false
          description: 검색할 혜택 이름 키워드
          schema:
            type: string
            example: "할리스"
        - name: category
          in: query
          required: false
          description: 지도 페이지 내 카테고리 검색
          schema:
            type: string
            example: "푸드"
      responses:
        '200':
          description: 즐겨찾기 검색 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: FAVORITE_BENEFIT_SEARCH_SUCCESS
                status: OK
                message: 즐겨찾기 검색이 완료되었습니다
                data:
                  - benefitId: 1
                    benefitName: "할리스"
                    partnerName: "할리스"
                    partnerImage: "https://itplacepartners.s3.ap-northeast-2.amazonaws.com/img/hollys.png"
                timestamp: '2025-07-15T02:20:54.631Z'

        '404':
          description: 사용자 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: USER_NOT_FOUND
                status: NOT_FOUND
                message: 사용자를 찾을 수 없습니다.
                data: null
                timestamp: '2025-07-16T15:00:00.000Z'
  /favorites/benefits/{benefitId}:
    get:
      tags:
        - Favorites
      summary: 즐겨찾기 혜택 상세 조회
      description: 혜택 ID를 기반으로 상세 정보를 조회합니다.
      operationId: getBenefitDetail
      parameters:
        - name: benefitId
          in: path
          required: true
          description: 혜택 ID
          schema:
            type: integer
      responses:
        '200':
          description: 혜택 상세 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: FAVORITE_BENEFIT_DETAIL_SUCCESS
                status: OK
                message: 혜택 상세 정보 조회 성공
                data:
                  benefitId: 3
                  benefitName: "CGV"
                  benefitDescription: "영화 그 이상의 감동 CGV"
                  benefitLimit: "VVIP/VIP 등급 전용\r VIP콕 내 무료예매 연3회/1+1예매 연9회(총 12회)\r (월 1회 사용 가능, CGV/메가박스 중 택 1)"
                  partnerName: "CGV"
                  partnerImage: "https://itplacepartners.s3.ap-northeast-2.amazonaws.com/img/cgv.png"
                  tiers:
                    - grade: VIP콕
                      isAll: false
                      context: "무료 예매 연 3회 / 1+1 예매 연 9회(총 12회)\r\n(월 1회 사용 가능, CGV/메가박스 중 택 1)\r\n※ 1+1이란? 1매 유료로 구매 시, 1매 무료 제공"
                      discountValue: 14000
                timestamp: '2025-07-15T12:00:00Z'

        '404':
          description: 혜택 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                code: BENEFIT_NOT_FOUND
                status: NOT_FOUND
                message: 존재하지 않는 혜택입니다.
                data: null
                timestamp: '2025-07-16T15:00:00.000Z'

  # =====================================================
  # User 관련
  # =====================================================
...
  /benefit/{benefitId}:
    get:
      tags:
        - Benefit
      summary: 혜택 상세 조회
      description: 혜택 ID를 통해 혜택의 상세 정보를 조회합니다.
      operationId: getBenefitDetail
      parameters:
        - name: benefitId
          in: path
          required: true
          description: 조회할 혜택의 ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 혜택 상세 조회 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/BenefitDetailResponse'
              example:
                data:
                  benefitId: 26
                  benefitName: "딸기콩"
                  description: "딸기콩은 웅진씽크빅에서 만든 국내 최대 어린이 오디오북 플랫폼으로 7,000여 개의 어린이 전문 콘텐츠를 기반으로 아이의 연령과 관심 영역에 맞춘 큐레이션을 제공합니다."
                  benefitLimit: "VIP콕 내 제휴사 통합 월 1회, 연 12회 가능"
                  manual: "멤버십앱 VIP 특별혜택 > 혜택 > 딸기콩 > 하단 '바로가기' 클릭 > 딸기콩 연동 후 등록\n\n\n\n\n*꼭 확인하세요\n\n\n- 딸기콩 구독권 또는 쿠폰 보유 시, 혜택 적용이 불가합니다.\n\n- 딸기콩 구독권 또는 쿠폰을 보유하고 계실 경우 딸기콩 1:1 문의 또는 웅진 고객센터(1577-1500)를 통해 혜택 적용이 가능합니다.\n\n- 본 혜택은 U+멤버십 혜택이 매달 자동 적용되지 않으며, 고객님께서 해당 페이지를 통해 매달 직접 신청하셔야 합니다.\n\n- 만 19세 이상 고객만 가입할 수 있습니다. \n\n\n\n대표문의\n\n딸기콩 고객센터 1577-1500"
                  url: "https://www.ddalgicong.com/login/login.html"
                  partnerName: "딸기콩"
                  image: ""
                code: BENEFIT_DETAIL_SUCCESS
                status: OK
                message: 혜택 상세 조회에 성공했습니다.
                timestamp: '2025-07-16T12:00:00.000Z'

        '404':
          description: 혜택을 찾을 수 없습니다
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                data: null
                code: BENEFIT_NOT_FOUND
                status: NOT_FOUND
                message: 존재하지 않는 혜택입니다.
                timestamp: '2025-07-16T12:00:00.000Z'
...
components:
  schemas:
...
    ApiResponse:
      type: object
      properties:
        code:
          type: string
          description: "서버에서 정의한 응답 코드"
          example: "SUCCESS_200_1"
        status:
          type: string
          description: "HTTP 상태 메시지"
          example: "OK"
        message:
          type: string
          description: "응답에 대한 설명 메시지"
          example: "요청이 정상적으로 처리되었습니다."
        data:
          type: object
          description: "실제 응답 데이터 (nullable)"
          nullable: true
        timestamp:
          type: string
          format: date-time
          description: "응답 시간"

    # ----------------------------------------------------
    # Request DTOs
    # ----------------------------------------------------
...
    FavoriteRequest:
      type: object
      properties:
        benefitId:
          type: integer
          example: 100
    RemoveFavoritesRequest:
      type: object
      properties:
        benefitIds:
          type: array
          items:
            type: integer
            example: [ 100, 101, 102 ]
    FavoriteResponse:
      type: object
      properties:
        benefitId:
          type: integer
          example: 100
        benefitName:
          type: string
          example: "CGV"
        partnerName:
          type: string
          example: "CGV"
        partnerImage:
          type: string
          example: "https://example.com/image.png"
...
    PageResult:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/FavoriteResponse'
        totalElements:
          type: integer
        totalPages:
          type: integer
        currentPage:
          type: integer
        hasNext:
          type: boolean
...
    BenefitDetailResponse:
      type: object
      properties:
        benefitId:
          type: integer
          format: int64
        benefitName:
          type: string
        description:
          type: string
        benefitLimit:
          type: string
        manual:
          type: string
        url:
          type: string
          format: uri
        partnerName:
          type: string
        image:
          type: string
          format: uri
...